// frontend/src/company.d.ts
export interface CompanyKeyMetrics {
  revenuePerShareTTM: number
  netIncomePerShareTTM: number
  operatingCashFlowPerShareTTM: number
  freeCashFlowPerShareTTM: number
  cashPerShareTTM: number
  bookValuePerShareTTM: number
  tangibleBookValuePerShareTTM: number
  shareholdersEquityPerShareTTM: number
  interestDebtPerShareTTM: number
  marketCapTTM: number
  enterpriseValueTTM: number
  peRatioTTM: number
  priceToSalesRatioTTM: number
  pocfratioTTM: number
  pfcfRatioTTM: number
  pbRatioTTM: number
  ptbRatioTTM: number
  evToSalesTTM: number
  enterpriseValueOverEBITDATTM: number
  evToOperatingCashFlowTTM: number
  evToFreeCashFlowTTM: number
  earningsYieldTTM: number
  freeCashFlowYieldTTM: number
  debtToEquityTTM: number
  debtToAssetsTTM: number
  netDebtToEBITDATTM: number
  currentRatioTTM: number
  interestCoverageTTM: number
  incomeQualityTTM: number
  dividendYieldTTM: number
  dividendYieldPercentageTTM: number
  payoutRatioTTM: number
  salesGeneralAndAdministrativeToRevenueTTM: number
  researchAndDevelopementToRevenueTTM: number
  intangiblesToTotalAssetsTTM: number
  capexToOperatingCashFlowTTM: number
  capexToRevenueTTM: number
  capexToDepreciationTTM: number
  stockBasedCompensationToRevenueTTM: number
  grahamNumberTTM: number
  roicTTM: number
  returnOnTangibleAssetsTTM: number
  grahamNetNetTTM: number
  workingCapitalTTM: number
  tangibleAssetValueTTM: number
  netCurrentAssetValueTTM: number
  investedCapitalTTM: number
  averageReceivablesTTM: number
  averagePayablesTTM: number
  averageInventoryTTM: number
  daysSalesOutstandingTTM: number
  daysPayablesOutstandingTTM: number
  daysOfInventoryOnHandTTM: number
  receivablesTurnoverTTM: number
  payablesTurnoverTTM: number
  inventoryTurnoverTTM: number
  roeTTM: number
  capexPerShareTTM: number
  dividendPerShareTTM: number
  debtToMarketCapTTM: number
}